Moved classes to tech.pegasys.web3signer.common package
Mainly to avoid a circular dependency between core and common packages
