+++ common/src/main/java/tech/pegasys/web3signer/common/PlatformDetector.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.web3signer.commandline;
+package tech.pegasys.web3signer.common;
 
 import java.util.Locale;
